Update firebox.bat
Changed the text back to "Enter password", changed the secure-password text back, since it was used to the creation as well, then removed it from the setup because it is actually handy to know if the password you enter is correct before creating and later on realizing you had a typo in it. Also removed the check for getpwd.ps1 because it is actually safer to always overwrite the file before using it in this context (another safety measure).